backend/requirements.txt
requests
python-dotenv
networkx
fastapi
uvicorn